requirements.txt
torch
torchvision
transformers
Pillow
numpy
pandas
tqdm